Consolidate ActiveClassChip tests into tests/components/common
Removes the colocated duplicate at components/common/ActiveClassChip.test.tsx
and merges its focus-management coverage into the canonical
tests/components/common/ActiveClassChip.test.tsx so the suite isn't
registered twice and matches the rest of the components/common test
layout.
